move checks for the zoom level to the front of the expressions
This should make rendering at lower zoom levels much more efficient, as there
are many objects per tile and most expressions will probably not match because
of the zoom level. The check for the zoom level is relatively cheap as it is
just comparing two integers, so it avoids more expensive checks like string
equality or regular expression matches when those objects will not be drawn
anyway. On the other hand this makes rendering on lower zoom levels slightly
less efficient.