fix JavaScript functions being added to value_tags map
In case the value of a text element is calculated by a function (e.g. join())
the JavaScript code that did the calculation was added to the value_tags map:

  MapCSS.e_join(" ", MapCSS.e_tag(tags, "razed:ref"), MapCSS.e_tag(tags, "razed:name"))

Check if the expression comes from an eval statement, and simply ignore it in
that case. Instead add the tags to the map when the call to MapCSS.e_tag() is
generated.