use different operators when comparing with numbers
When checking a value against a number in case of equality currently the
operator '==' was used (equality without type checking), but in case of
inequality the operator '!==' was used (inequality with type checking). Since
the value checked was always a string and the other value was a number the type
was never the same and the inequality operator always returned true.

From now on the type equality is never used when comparing to a number, thus
making it basically work as expected. When comparing other types (i.e. strings
on both sides) always check for type equality to avoid any conversions from
happening.