fix more rendering artifacts by avoiding overly long lines
Until now the "draw beyond tile border" offset was applied in screen
coordinates. This offset was 50, i.e. 10% of a tile. If both ends of the line
this was applied were on a tile boundary this could have been doubled, i.e. one
point got 50px offset to one direction and the other got 50px to the other one,
leading in a virtual line of 100px.

Now it could happen that the line in question was aligned relatively close to
the tile border, i.e. the angle between the tile border and the line was nearly
zero. In that case the rendering artifact was visible as a horizontal or
vertical line along the tile border.

The code is now changed to apply that line prolonging in tile coordinates, i.e.
before converting it to screen coordinates. This reduces the impact of from 10%
to 0.5%, as a vector tile has coordinates from 0 to 10000. This should still be
long enough to avoid broken lines at that point.
 In case those line prolonging was more or
less along the tile borders