support additional font properties
This makes the properties font-weight, font-style, and font-variant work as
defined. Until now only font-name and font-size were respected. For shields
the prefixed values of font-name and font-size are checked first, falling
back to the unprefixed values if not defined. The new properties are only
checked in their unprefixed version.