drop the "bold" and "italic" autodetection by font name
This could errnously trigger if a font accidentially had this as part of it's
own name (e.g. "Exambolda"). Since font-weight and font-style properties are
supported now simply require the callers to properly set those properties and
not rely on the font name matching.

It also does not add the non-bold font variant to the font list automatically
anymore. If that one should also be used the user can simply specify it.