try different text wrapping methods until a label can be drawn
Try not only splitting the label on different points, but also the unmodified
string. This make long station names shorter so it's more likely for them to
show up on the map as they can now avoid to collide with vertical tile borders.

While at it calculate the line height using a fixed factor instead of the mean
character width of a line which will give a consistent look now that this is not
only used for collision calculation, but is also user visible.