fix rendering of polygons where one border goes from one tile boundary to another
Think of a square that is exactly the same size as the tile, but rotated by 45
degrees. Now the object in this tile consists of 8 points, 2 of them on each
tile boundary. No line should be drawn between each of those pairs, because
those are in fact no outer borders of the square. But the other lines were not
drawn either because it was only checked if the points were on tile boundaries,
but not on which ones.

From now on lines between points on tile borders will be drawn if the points are
on distinct borders.